+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+- Always increment build number
+- Added `no-build-increment` flag to prevent incrementing the build number
+- Added `version` flag to specify the version number to set
+
 ## 1.0.1
 - Added `-b` flag to also increment the build number